# api/openapi.yaml
openapi: 3.0.3
info:
  title: Arc Issue Tracker API
  description: |
    Central issue tracking server for AI-assisted coding workflows.
    Arc provides project-scoped issue management with dependency tracking,
    labels, comments, and audit trails.
  version: 1.0.0
  contact:
    name: SentioLabs
    url: https://github.com/sentiolabs/arc
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /projects/{projectId}/issues:
    parameters:
      - $ref: "#/components/parameters/ProjectId"

    get:
      operationId: listIssues
      tags: [issues]
      summary: List issues in project
      parameters:
        - name: status
          in: query
          description: Filter by status
          schema:
            $ref: "#/components/schemas/Status"
        - name: type
          in: query
          description: Filter by issue type
          schema:
            $ref: "#/components/schemas/IssueType"
        - name: priority
          in: query
          description: Filter by priority (0-4)
          schema:
            type: integer
            minimum: 0
            maximum: 4
        - name: assignee
          in: query
          description: Filter by assignee
          schema:
            type: string
        - name: parent_id
          in: query
          required: false
          description: Filter by parent issue ID (returns children via parent-child dependency)
          schema:
            type: string
        - name: q
          in: query
          description: Full-text search query
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum results to return
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
        - name: offset
          in: query
          description: Pagination offset
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: Paginated list of issues
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedIssues"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: createIssue
      tags: [issues]
      summary: Create a new issue
      parameters:
        - $ref: "#/components/parameters/ActorHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIssueRequest"
      responses:
        "201":
          description: Issue created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /projects/{projectId}/issues/{issueId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/IssueId"

    get:
      operationId: getIssue
      tags: [issues]
      summary: Get issue by ID
      parameters:
        - name: details
          in: query
          description: Include full details (dependencies, comments, labels)
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Issue details
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Issue"
                  - $ref: "#/components/schemas/IssueDetails"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      operationId: updateIssue
      tags: [issues]
      summary: Update issue
      parameters:
        - $ref: "#/components/parameters/ActorHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIssueRequest"
      responses:
        "200":
          description: Issue updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      operationId: deleteIssue
      tags: [issues]
      summary: Delete issue
      responses:
        "204":
          description: Issue deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /projects/{projectId}/issues/{issueId}/close:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/IssueId"

    post:
      operationId: closeIssue
      tags: [issues]
      summary: Close an issue
      parameters:
        - $ref: "#/components/parameters/ActorHeader"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloseIssueRequest"
      responses:
        "200":
          description: Issue closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /projects/{projectId}/issues/{issueId}/reopen:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/IssueId"

    post:
      operationId: reopenIssue
      tags: [issues]
      summary: Reopen a closed issue
      parameters:
        - $ref: "#/components/parameters/ActorHeader"
      responses:
        "200":
          description: Issue reopened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ====================
  # Ready & Blocked
  # ====================
...
  /projects/{projectId}/blocked:
    parameters:
      - $ref: "#/components/parameters/ProjectId"

    get:
      operationId: getBlockedIssues
      tags: [ready]
      summary: Get blocked issues
      parameters:
        - name: limit
          in: query
          description: Maximum results to return
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
      responses:
        "200":
          description: List of blocked issues with blocking info
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlockedIssue"
        "500":
          $ref: "#/components/responses/InternalError"

  # ====================
  # Team Context
  # ====================
...
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      description: Project ID
      schema:
        type: string
    IssueId:
      name: issueId
      in: path
      required: true
      description: Issue ID
      schema:
        type: string
    ActorHeader:
      name: X-Actor
      in: header
      description: User performing the action (defaults to "anonymous")
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
    Status:
      type: string
      enum:
        - open
        - in_progress
        - blocked
        - deferred
        - closed
    IssueType:
      type: string
      enum:
        - bug
        - feature
        - task
        - epic
        - chore
    DependencyType:
      type: string
      enum:
        - blocks
        - parent-child
        - related
        - discovered-from
...
    Issue:
      type: object
      required:
        - id
        - project_id
        - title
        - status
        - priority
        - issue_type
        - created_at
        - updated_at
      properties:
        id:
          type: string
          description: Unique issue ID
        project_id:
          type: string
        title:
          type: string
          maxLength: 500
        description:
          type: string
        status:
          $ref: "#/components/schemas/Status"
        priority:
          type: integer
          minimum: 0
          maximum: 4
          description: "0 (critical) - 4 (backlog)"
        rank:
          type: integer
          description: "0 = unranked (sorts last), 1+ = lower rank = work on first"
        issue_type:
          $ref: "#/components/schemas/IssueType"
        assignee:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        closed_at:
          type: string
          format: date-time
        close_reason:
          type: string
        external_ref:
          type: string
          description: External reference (e.g., "gh-9", "jira-ABC")
        labels:
          type: array
          items:
            type: string
        dependencies:
          type: array
          items:
            $ref: "#/components/schemas/Dependency"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
    IssueDetails:
      allOf:
        - $ref: "#/components/schemas/Issue"
        - type: object
          properties:
            dependents:
              type: array
              items:
                $ref: "#/components/schemas/Dependency"
    BlockedIssue:
      allOf:
        - $ref: "#/components/schemas/Issue"
        - type: object
          required:
            - blocked_by_count
            - blocked_by
          properties:
            blocked_by_count:
              type: integer
            blocked_by:
              type: array
              items:
                type: string
    PaginatedIssues:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Issue"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    # ====================
    # Team Context Schemas
    # ====================
...
    CreateIssueRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 500
        description:
          type: string
        status:
          $ref: "#/components/schemas/Status"
        priority:
          type: integer
          minimum: 0
          maximum: 4
          default: 2
        issue_type:
          $ref: "#/components/schemas/IssueType"
        assignee:
          type: string
        external_ref:
          type: string
    UpdateIssueRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 500
        description:
          type: string
        status:
          $ref: "#/components/schemas/Status"
        priority:
          type: integer
          minimum: 0
          maximum: 4
        issue_type:
          $ref: "#/components/schemas/IssueType"
        assignee:
          type: string
        external_ref:
          type: string
    CloseIssueRequest:
      type: object
      properties:
        reason:
          type: string
          description: Reason for closing

    # ====================
    # Dependency Schemas
    # ====================
    Dependency:
      type: object
      required:
        - issue_id
        - depends_on_id
        - type
        - created_at
      properties:
        issue_id:
          type: string
        depends_on_id:
          type: string
        type:
          $ref: "#/components/schemas/DependencyType"
        created_at:
          type: string
          format: date-time
        created_by:
          type: string
...
    Comment:
      type: object
      required:
        - id
        - issue_id
        - author
        - text
        - created_at
      properties:
        id:
          type: integer
          format: int64
        issue_id:
          type: string
        author:
          type: string
        text:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...